Google_ADK/7-multi-agents/requirements.txt
google-adk
yfinance
litellm